testing on linux

+++ tests/queue_test.c
@@ -11,7 +11,7 @@ static int queue_length(const struct queue *queue)
     return i;
 }
 
-static void queue_to_array(const struct queue *queue, long int *array)
+static void queue_to_array(const struct queue *queue, size_t *array)
 {
     int i = 0;
     for (struct linked_list *link = queue->front; link; link = link->next) {
@@ -22,7 +22,7 @@ static void queue_to_array(const struct queue *queue, long int *array)
 int main(int argc, char * argv[])
 {
     struct queue *queue = empty_queue();
-    int array[10];
+    size_t array[10];
     
     assert(queue_length(queue) == 0);
     enqueue(queue, (void *)1);
